Module Nfa.Trans

Functor for converting a NFA, with a given implementation of state identifiers and symbols into another one with different respective implementations

Parameters

module S1 : T

module S2 : T

Signature

val map : (S1.state -> S2.state) -> (S1.symbol -> S2.symbol) -> S1.t -> S2.t