Module Lascar__Ltsa

Labeled Transition Systems with State Attributes

A Labeled Transition System with Attributes on states (LTSA) is 6-uple (Q,L,A,I,a,R) where

Q is a set of state identifiers

L is a set of transition labels

A is a set of state attributes

I is a subset of Q giving the initial states

a is function from Q to A associating a state to its attribute

R is the transition relation, defined as a subset of QxLxQ

type Utils.Dot.graph_style += Circular | NoAttr | NoLabel

Extension to the Dot.graph_style type for drawing LTSAs

module type STATE = Utils.OrderedTypeExt.T

Signature of the module describing state identifiers

module type LABEL = Utils.OrderedTypeExt.T

Signature of the module describing transition labels

module type ATTR = sig ... end

Signature of the module describing state attributes

module type T = sig ... end

Output signature of the functor Ltsa.Make.

module Make : functor (S : STATE) -> functor (L : LABEL) -> functor (A : ATTR) -> T with type state = S.t and type label = L.t and type attr = A.t

module Trans : functor (S1 : T) -> functor (S2 : T) -> sig ... end

module Product : functor (S1 : T) -> functor (S2 : T) -> sig ... end

Functor for computing the external product, in different flavors, of two LTSA

module Product3 : functor (S1 : T) -> functor (S2 : T) -> functor (S3 : T) -> sig ... end

Functor for computing the external product, in different flavors, of three LTSA

module type Merge = sig ... end

Signature of the second argument for the IProduct functor

module IProduct : functor (S : T) -> functor (M : Merge with type state = S.state and type label = S.label and type attr = S.attr) -> sig ... end

Functor for computing the internal product, in different flavors, of two LTSA

module type Merge3 = sig ... end

Signature of the second argument for the IProduct3 functor

module IProduct3 : functor (S : T) -> functor (M : Merge3 with type state = S.state and type label = S.label and type attr = S.attr) -> sig ... end

Functor for computing the "internal" product, in different flavors, of three LTSA