Module Utils__SetExt

Extension to the Stdib.Set module

module type S = sig ... end

Output signature of the SetExt.Make functor.

module Make : functor (E : Utils.OrderedTypeExt.T) -> S with type Make.elt = E.t

Functor building an implementation of the set structure given a totally ordered, stringable type.

module Product : functor (S1 : S) -> functor (S2 : S) -> sig ... end

Functor building an implementation of a set product given the implementation of two sets