Parameter Make.1-A

type node

The type of information attached to the nodes of the tree

type edge

The type of information attached to the nodes of the tree

The type of information attached to the edges of the tree

val string_of_node : node -> string

The type of information attached to the edges of the tree

val string_of_edge : edge -> string