Module Lascar.Conv

Functors for converting various kinds of Labeled Transition Systems

module ToLts (M : Ltsa.T) : sig ... end

Functor for converting a Ltsa into a Lts (by removing state attributes)

module FromLts (M : Lts.T) : sig ... end

Functor for converting a Lts into a Ltsa (by adding unit state attributes)

module ToDfa (N : Nfa.T) : sig ... end

Functor for converting a Nfa to an equivalent Dfa (determinisation)

module ToMealy (MM : Moore.T) : sig ... end

Functor for converting a Moore machine into an equivalent Mealy one

module ToMoore (ME : Mealy.T) : sig ... end

Functor for converting a Mealy machine into an equivalent Moore one

module Fsm (F : Fsm.T) : sig ... end

Functor for transforming FSMs