Module Lascar.Dfa

Deterministic Finite Automata (DFA)

A DFA is just a NFA in which there's at most one transition with a given symbol from any state. This condition will be enforced by the create and add_transition functions.

module type T = sig ... end

module Make (S : Ltsa.STATE) (L : Nfa.SYMBOL) : T with type state = S.t and type symbol = L.t

Functor building an implementation of the DFA structure given an implementation of state identifiers and symbols

module Trans (S1 : T) (S2 : T) : sig ... end

Functor for converting a DFA, with a given implementation of state identifiers and symbols into another one with different respective implementations

module Product (S1 : T) (S2 : T with type symbol = S1.symbol and type Symbols.t = S1.Symbols.t and type NFA.Symbols.t = S1.NFA.Symbols.t) : sig ... end

Functor for computing the products of two NFAs sharing the same symbol set