Module Utils.FilenameExt

Extension to the Stdlib.Filename module

val split_suffix : string -> string * string

split_suffix fname removes any suffix from fname and returns both the resulting name and the removed suffix

val add_before_suffix : string -> string -> string

if split_suffix fname gives (base,suff), add_before_suffix fname s returns base^s^suff

val replace_suffix : string -> string -> string

if split_suffix fname gives (base,suff), replace_suffix fname s returns base^"."^s